@@ -1,10 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk" ToolsVersion="Current">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <FileUpgradeFlags>40</FileUpgradeFlags>
     <UpgradeBackupLocation>C:\Users\Beebster\Documents\Source\Repos\EQOA_Proto_C-sharp\Backup1\</UpgradeBackupLocation>
     <OldToolsVersion>Current</OldToolsVersion>
+    <AnalysisLevel>latest</AnalysisLevel>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
@@ -31,9 +32,32 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="BCrypt.Net-Next" Version="4.0.0" />
+    <PackageReference Include="Microsoft.Extensions.ObjectPool" Version="6.0.4" />
     <PackageReference Include="MySqlConnector" Version="0.69.4" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="NLua" Version="1.6.0" />
     <PackageReference Include="QuadTrees" Version="1.0.4" />
     <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.7.0" />
     <PackageReference Include="System.IO.Pipelines" Version="5.0.1" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="$(ProjectDir)\Scripts\**\*;$(ProjectDir)\Scripts\*">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Remove="Scripts\wyndhaven\coachman_brumley.lua" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Remove="Scripts\ports.lua" />
+    <None Remove="Scripts\wyndhaven\coachman_brumley.lua" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="Scripts\dirTest.lua" />
+    <Content Include="Scripts\ports.lua" />
+    <Content Include="Scripts\wyndhaven\coachman_brumley.lua" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Scripts\forkwatch\" />
   </ItemGroup>
 </Project>
