@@ -9,14 +9,16 @@ namespace ReturnHome.Server.EntityObject.Player
     public class Item
     {
         //Should these be public? How would we handle adding stats to character overall stats without 100 methods?
-        public int StackLeft { get; private set; }
-        public int RemainingHP { get; private set; }
-        public int Charges { get; private set; }
-        public int EquipLocation { get; private set; }
-        public byte Location { get; private set; } //inventory, bank auction etc
-        public int InventoryNumber { get; private set; } //Location in inventory, would location in List suffice for this?
+        public byte ClientIndex;
+        public int StackLeft { get; set; }
+        public int RemainingHP { get; set; }
+        public int Charges { get; set; }
+        public int EquipLocation { get; set; }
+        public byte Location { get; set; } //inventory, bank auction etc
+        public int ServerKey { get; set; } //Location in inventory, would location in List suffice for this?
+
         public int ItemID { get; private set; }
-        public int ItemCost { get; private set; }
+        public uint ItemCost { get; private set; }
         public int Unk1 { get; private set; }
         public int ItemIcon { get; private set; }
         public int Unk2 { get; private set; }
@@ -73,12 +75,13 @@ namespace ReturnHome.Server.EntityObject.Player
         //This will instantiate an inventory object
         //Should we be able to instantiate a normal item and gear seperately? Seems the better choice
         //This is to instantiate 
-        public Item(int thisStacksLeft, int thisCharges, byte thisLocation, int thisInventoryNumber, int thisItemID, int thisItemCost, int thisItemIcon, int thisTrade, int thisRent, int thisCraft, int thisLore, int thisLevelreq, int thisMaxStack, string thisItemName, string thisItemDesc)
+        public Item(int thisStacksLeft, int thisCharges, byte thisLocation, byte thisInventoryNumber, int thisItemID, uint thisItemCost, int thisItemIcon, int thisTrade, int thisRent, int thisCraft, int thisLore, int thisLevelreq, int thisMaxStack, string thisItemName, string thisItemDesc)
+
         {
             StackLeft = thisStacksLeft;
             Charges = thisCharges;
             Location = thisLocation;
-            InventoryNumber = thisInventoryNumber;
+            ServerKey = thisInventoryNumber;
             ItemID = thisItemID;
             ItemCost = thisItemCost;
             ItemIcon = thisItemIcon;
@@ -96,12 +99,13 @@ namespace ReturnHome.Server.EntityObject.Player
         //Constructor object for armour and weapons
         //Alot of this could be managed by scripting as there is a huge portion that is static
         //Varis: int thisStacksLeft, int thisRemainingHP, int thisCharges, int thisEquipLocation, byte thisLocation, int thisInventoryNumber, int thisItemID<- use this in scripting to get right gear?
-        public Item(int thisStacksLeft, int thisRemainingHP, int thisCharges, int thisEquipLocation, byte thisLocation, int thisInventoryNumber, int thisItemID, int thisItemCost, int thisItemIcon, int thisEquipslot, int thisAttackType, int thisWeaponDamage, int thisMaxHP, int thisTrade, int thisRent, int thisCraft, int thisLore, int thisLevelreq, int thisMaxStack, string thisItemName, string thisItemDesc, int thisDuration, int thisClassuse, int thisRaceuse, int thisProcanim, int Strength, int Stamina, int Agility, int Dexterity, int Wisdom, int Intelligence, int Charisma, int HpMax, int PowMax, int pot, int hot, int ac, int pr, int dr, int fr, int cr, int lr, int ar, int model, uint color)
+        public Item(int thisStacksLeft, int thisRemainingHP, int thisCharges, int thisEquipLocation, byte thisLocation, byte thisInventoryNumber, int thisItemID, uint thisItemCost, int thisItemIcon, int thisEquipslot, int thisAttackType, int thisWeaponDamage, int thisMaxHP, int thisTrade, int thisRent, int thisCraft, int thisLore, int thisLevelreq, int thisMaxStack, string thisItemName, string thisItemDesc, int thisDuration, int thisClassuse, int thisRaceuse, int thisProcanim, int Strength, int Stamina, int Agility, int Dexterity, int Wisdom, int Intelligence, int Charisma, int HpMax, int PowMax, int pot, int hot, int ac, int pr, int dr, int fr, int cr, int lr, int ar, int model, uint color)
+
         {
             StackLeft = thisStacksLeft;
             Charges = thisCharges;
             Location = thisLocation;
-            InventoryNumber = thisInventoryNumber;
+            ServerKey = thisInventoryNumber;
             ItemID = thisItemID;
             ItemCost = thisItemCost;
             ItemIcon = thisItemIcon;
@@ -150,187 +154,194 @@ namespace ReturnHome.Server.EntityObject.Player
             AR = ar;
         }
 
-        public void DumpItem(MemoryStream memStream)
+        public Item AcquireItem(int qty)
         {
-            //Start adding attributes to list for this item
-            memStream.Write(Utility_Funcs.DoublePack(StackLeft));
-            memStream.Write(Utility_Funcs.DoublePack(RemainingHP));
-            memStream.Write(Utility_Funcs.DoublePack(Charges));
-            memStream.Write(Utility_Funcs.DoublePack(EquipLocation));
-            memStream.WriteByte(Location);
-            memStream.Write(BitConverter.GetBytes(InventoryNumber));
-            memStream.Write(Utility_Funcs.DoublePack(ItemID));
-            memStream.Write(Utility_Funcs.DoublePack(ItemCost));
-            memStream.Write(Utility_Funcs.DoublePack(Unk1));
-            memStream.Write(Utility_Funcs.DoublePack(ItemIcon));
-            memStream.Write(Utility_Funcs.DoublePack(Unk2));
-            memStream.Write(Utility_Funcs.DoublePack(Equipslot));
-            memStream.Write(Utility_Funcs.DoublePack(Unk3));
-            memStream.Write(Utility_Funcs.DoublePack(Trade));
-            memStream.Write(Utility_Funcs.DoublePack(Rent));
-            memStream.Write(Utility_Funcs.DoublePack(Unk4));
-            memStream.Write(Utility_Funcs.DoublePack(Attacktype));
-            memStream.Write(Utility_Funcs.DoublePack(Weapondamage));
-            memStream.Write(Utility_Funcs.DoublePack(Unk5));
-            memStream.Write(Utility_Funcs.DoublePack(Levelreq));
-            memStream.Write(Utility_Funcs.DoublePack(Maxstack));
-            memStream.Write(Utility_Funcs.DoublePack(Maxhp));
-            memStream.Write(Utility_Funcs.DoublePack(Duration));
-            memStream.Write(Utility_Funcs.DoublePack(Classuse));
-            memStream.Write(Utility_Funcs.DoublePack(Raceuse));
-            memStream.Write(Utility_Funcs.DoublePack(Procanim));
-            memStream.Write(Utility_Funcs.DoublePack(Lore));
-            memStream.Write(Utility_Funcs.DoublePack(Unk6));
-            memStream.Write(Utility_Funcs.DoublePack(Craft));
-            memStream.Write(BitConverter.GetBytes(ItemName.Length));
-            memStream.Write(Encoding.Unicode.GetBytes(ItemName));
-            memStream.Write(BitConverter.GetBytes(ItemDesc.Length));
-            memStream.Write(Encoding.Unicode.GetBytes(ItemDesc));
-            PullStats(memStream);
+            Item item = (Item)MemberwiseClone();
+            item.StackLeft = qty;
+            return item;
         }
 
-        private void PullStats(MemoryStream memStream)
+        public void DumpItem(ref BufferWriter writer)
+        {
+            //Start adding attributes to list for this item
+            writer.Write7BitEncodedInt64(StackLeft);
+            writer.Write7BitEncodedInt64(RemainingHP);
+            writer.Write7BitEncodedInt64(Charges);
+            writer.Write7BitEncodedInt64(EquipLocation);
+            writer.Write(Location);
+            writer.Write(ServerKey);
+            writer.Write7BitEncodedInt64(ItemID);
+            writer.Write7BitEncodedUInt64(ItemCost);
+            writer.Write7BitEncodedInt64(Unk1);
+            writer.Write7BitEncodedInt64(ItemIcon);
+            writer.Write7BitEncodedInt64(Unk2);
+            writer.Write7BitEncodedInt64(Equipslot);
+            writer.Write7BitEncodedInt64(Unk3);
+            writer.Write7BitEncodedInt64(Trade);
+            writer.Write7BitEncodedInt64(Rent);
+            writer.Write7BitEncodedInt64(Unk4);
+            writer.Write7BitEncodedInt64(Attacktype);
+            writer.Write7BitEncodedInt64(Weapondamage);
+            writer.Write7BitEncodedInt64(Unk5);
+            writer.Write7BitEncodedInt64(Levelreq);
+            writer.Write7BitEncodedInt64(Maxstack);
+            writer.Write7BitEncodedInt64(Maxhp);
+            writer.Write7BitEncodedInt64(Duration);
+            writer.Write7BitEncodedInt64(Classuse);
+            writer.Write7BitEncodedInt64(Raceuse);
+            writer.Write7BitEncodedInt64(Procanim);
+            writer.Write7BitEncodedInt64(Lore);
+            writer.Write7BitEncodedInt64(Unk6);
+            writer.Write7BitEncodedInt64(Craft);
+            writer.WriteString(Encoding.Unicode, ItemName);
+            writer.WriteString(Encoding.Unicode, ItemDesc);
+            PullStats(ref writer);
+        }
+
+        private void PullStats(ref BufferWriter writer)
         {
 
             //Gather stats if any exist
             //Increment counter if if statement true, then add identifier int for stat and then DoublePack and add stat value
-            long position = memStream.Position;
+            int position = writer.Position;
             //Place 0 place holder here
-            memStream.WriteByte(0);
+            writer.Write((byte)0);
 
             if (Str > 0)
             {
                 Counter++;
-                memStream.WriteByte(0);
-                memStream.Write(Utility_Funcs.DoublePack(Str));
+                writer.Write((byte)0);
+                writer.Write7BitEncodedInt64(Str);
             }
 
             if (Sta > 0)
             {
                 Counter++;
-                memStream.WriteByte(2);
-                memStream.Write(Utility_Funcs.DoublePack(Sta));
+                writer.Write((byte)2);
+                writer.Write7BitEncodedInt64(Sta);
             }
 
             if (Agi > 0)
             {
                 Counter++;
-                memStream.WriteByte(4);
-                memStream.Write(Utility_Funcs.DoublePack(Agi));
+                writer.Write((byte)4);
+                writer.Write7BitEncodedInt64(Agi);
             }
 
             if (Dex > 0)
             {
                 Counter++;
-                memStream.WriteByte(6);
-                memStream.Write(Utility_Funcs.DoublePack(Dex));
+                writer.Write((byte)6);
+                writer.Write7BitEncodedInt64(Dex);
             }
 
             if (Wis > 0)
             {
                 Counter++;
-                memStream.WriteByte(8);
-                memStream.Write(Utility_Funcs.DoublePack(Wis));
+                writer.Write((byte)8);
+                writer.Write7BitEncodedInt64(Wis);
             }
 
             if (Int > 0)
             {
                 Counter++;
-                memStream.WriteByte(10);
-                memStream.Write(Utility_Funcs.DoublePack(Int));
+                writer.Write((byte)10);
+                writer.Write7BitEncodedInt64(Int);
             }
 
             if (Cha > 0)
             {
                 Counter++;
-                memStream.WriteByte(12);
-                memStream.Write(Utility_Funcs.DoublePack(Cha));
+                writer.Write((byte)12);
+                writer.Write7BitEncodedInt64(Cha);
             }
 
             if (HPMax > 0)
             {
                 Counter++;
-                memStream.WriteByte(16);
-                memStream.Write(Utility_Funcs.DoublePack(HPMax));
+                writer.Write((byte)16);
+                writer.Write7BitEncodedInt64(HPMax);
             }
 
             if (POWMax > 0)
             {
                 Counter++;
-                memStream.WriteByte(20);
-                memStream.Write(Utility_Funcs.DoublePack(POWMax));
+                writer.Write((byte)20);
+                writer.Write7BitEncodedInt64(POWMax);
             }
 
             if (PoT > 0)
             {
                 Counter++;
-                memStream.WriteByte(24);
-                memStream.Write(Utility_Funcs.DoublePack(PoT));
+                writer.Write((byte)24);
+                writer.Write7BitEncodedInt64(PoT);
             }
 
             if (HoT > 0)
             {
                 Counter++;
-                memStream.WriteByte(26);
-                memStream.Write(Utility_Funcs.DoublePack(HoT));
+                writer.Write((byte)26);
+                writer.Write7BitEncodedInt64(HoT);
             }
 
             if (AC > 0)
             {
                 Counter++;
-                memStream.WriteByte(28);
-                memStream.Write(Utility_Funcs.DoublePack(AC));
+                writer.Write((byte)28);
+                writer.Write7BitEncodedInt64(AC);
             }
 
             if (PR > 0)
             {
                 Counter++;
-                memStream.WriteByte(44);
-                memStream.Write(Utility_Funcs.DoublePack(PR));
+                writer.Write((byte)44);
+                writer.Write7BitEncodedInt64(PR);
             }
 
             if (DR > 0)
             {
                 Counter++;
-                memStream.WriteByte(46);
-                memStream.Write(Utility_Funcs.DoublePack(DR));
+                writer.Write((byte)46);
+                writer.Write7BitEncodedInt64(DR);
             }
 
             if (FR > 0)
             {
                 Counter++;
-                memStream.WriteByte(48);
-                memStream.Write(Utility_Funcs.DoublePack(FR));
+                writer.Write((byte)48);
+                writer.Write7BitEncodedInt64(FR);
             }
 
             if (CR > 0)
             {
                 Counter++;
-                memStream.WriteByte(50);
-                memStream.Write(Utility_Funcs.DoublePack(CR));
+                writer.Write((byte)50);
+                writer.Write7BitEncodedInt64(CR);
             }
 
             if (LR > 0)
             {
                 Counter++;
-                memStream.WriteByte(52);
-                memStream.Write(Utility_Funcs.DoublePack(LR));
+                writer.Write((byte)52);
+                writer.Write7BitEncodedInt64(LR);
             }
 
             if (AR > 0)
             {
                 Counter++;
-                memStream.WriteByte(54);
-                memStream.Write(Utility_Funcs.DoublePack(AR));
+                writer.Write((byte)54);
+                writer.Write7BitEncodedInt64(AR);
             }
 
             if (Counter == 0)
                 return;
-            long position2 = memStream.Position;
-            memStream.Position = position;
-            memStream.Write(Utility_Funcs.DoublePack(Counter));
 
-            memStream.Position = position2;
+            int position2 = writer.Position;
+            writer.Position = position;
+            writer.Write7BitEncodedInt64(Counter);
+
+            writer.Position = position2;
+            Counter = 0;
         }
     }
 }
