@@ -1,17 +1,23 @@
 using System.Collections.Generic;
 using System.IO;
+using ReturnHome.Server.Network;
+using ReturnHome.Utilities;
 
 namespace ReturnHome.Server.EntityObject.Player
 {
     public class Quest
     {
-        private string ourMessage = "";
+        //private string ourMessage = "";
         public Quest()
         {
 
         }
 
-        public void DumpQuest(MemoryStream memStream)
+        public void DumpQuest(ref BufferWriter writer)
+        {
+
+        }
+        public static void AddQuestLog(Session MySession, Quest quest)
         {
 
         }
