@@ -2,6 +2,8 @@
 using System.Timers;
 using ReturnHome.Server.Managers;
 using ReturnHome.Server.Network.Managers;
+using ReturnHome.Server.EntityObject.Player;
+using ReturnHome.Server.Opcodes;
 
 namespace ReturnHome.Server
 {
@@ -29,8 +31,21 @@ namespace ReturnHome.Server
             // Start SocketManager
 			SocketManager.Initialize();
             Console.WriteLine("Server started.");
-			
+            /*EventManager eManager = new EventManager();
+            Dialogue newDiag = new Dialogue("Aeric Sparelli", 0, 0);
+            Character newCharacter = new Character();
+            newCharacter.MyDialogue = newDiag;
+            Dialogue thisDiag = eManager.GetNPCDialogue(GameOpcode.DialogueBox,newCharacter);
+            foreach(string dialogue in thisDiag.diagOptions)
+            {
+                Console.WriteLine(dialogue);
+            }*/
+           
+
             string stuff = Console.ReadLine();
+
+
+
         }
     }
 }
