@@ -4,11 +4,12 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Collections.Generic;
 using ReturnHome.Database.SQL;
-using ReturnHome.Server.EntityObject.Actor;
+using ReturnHome.Server.EntityObject.Actors;
 using ReturnHome.Server.EntityObject;
 using QuadTrees;
 
 using ReturnHome.Server.Network.Managers;
+using ReturnHome.Server.EntityObject.Player;
 
 namespace ReturnHome.Server.Managers
 {
@@ -16,7 +17,7 @@ namespace ReturnHome.Server.Managers
     {
 
         private static Stopwatch gameTimer;
-        private static int serverTick = 1000 / 5;
+        private static int serverTick = 1000 / 10;
 
         static WorldServer()
         {
@@ -25,30 +26,30 @@ namespace ReturnHome.Server.Managers
 
         public static void Initialize()
         {
-            var thread = new Thread(() =>
-                        {
-                            UpdateWorld();
-                        });
-            thread.Name = "World Manager";
-            thread.Priority = ThreadPriority.AboveNormal;
-            thread.Start();
-
             //Creates NPC List
             CharacterSQL npcList = new();
             //Calls sql query function that fills list full of NPCs
             List<Actor> myNpcList = npcList.WorldActors();
+            //Closing DB connection
+            npcList.CloseConnection();
+            MapManager.Initialize();
             //Loops through each npc in list and sets their position, adds them to the entity manager, and mapmanager
             Console.WriteLine("Adding NPCs...");
-            uint objectID = 0;
             foreach (Actor myActor in myNpcList)
             {
-                
-                
-                myActor.SetPosition();
                 EntityManager.AddEntity(myActor);
-                MapManager.AddObjectToTree(myActor);
+                MapManager.Add(myActor);
+
             }
             Console.WriteLine("Done.");
+
+            var thread = new Thread(() =>
+            {
+                UpdateWorld();
+            });
+            thread.Name = "World Manager";
+            thread.Priority = ThreadPriority.AboveNormal;
+            thread.Start();
         }
 
         public async static void UpdateWorld()
