@@ -5,11 +5,11 @@ namespace ReturnHome.Server.Network
 	//Contains bundle header flags along with BundleTypeFlags
     public enum PacketBundleFlags  : byte
     {
-        NewProcessMessages      = 0x00,
-		NewProcessReport        = 0x03,
-		ProcessMessageAndReport = 0x13,
-		ProcessMessages         = 0x20,
-		ProcessReport           = 0x23,
-		ProcessAll              = 0x63
+        NewProcessMessages        = 0x00,
+		RDPReport                 = 0x03,
+		ProcessUnreliableAcks     = 0x10,
+        UnknownSingleByte         = 0x08,
+        ProcessMessages           = 0x20,
+		ProcessAck                = 0x40
     }
 }
