@@ -1,7 +1,4 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 using ReturnHome.Utilities;
 using ReturnHome.Server.EntityObject;
@@ -14,91 +11,138 @@ namespace ReturnHome.Server.Network
         private Session _session;
 
         //Stores our base data to xor against
-        public Memory<byte> baseXOR = new Memory<byte>(new byte[0xC9]);
+        private Memory<byte> _baseXOR = new Memory<byte>(new byte[0xC9]);
 
         //This is when the client ack's a specific message, we then set this to that ack'd message # to help generate our xor
-        public ushort baseMessageCounter = 0;
+        private ushort _baseMessageCounter = 0;
+        public ushort BaseMessageCounter
+        {
+            get { return _baseMessageCounter; }
+            set
+            {
+                if (value > _baseMessageCounter)
+                    _baseMessageCounter = value;
+
+                else
+                    Console.WriteLine($"Error setting base message counter. Setting {value} Expected: {_baseMessageCounter}");
+            }
+        }
 
         //Simple message counter
-        public ushort messageCounter = 1;
+        private ushort _messageCounter = 1;
+        public ushort MessageCounter
+        {
+            get { return _messageCounter; }
+            set
+            {
+                if (value > _messageCounter)
+                    _messageCounter = value;
+
+                else
+                    Console.WriteLine($"Error setting message counter. Setting: {value} Expected: {_messageCounter}");
+            }
+        }
         //May not be needed
-        public byte ObjectChannel;
+        private byte _objectChannel;
 
         //Place to hold all of our XOR result's till client acks. We then xor that against the baseXOR to get new base object update, clear list once a message is ack'd by client
-        private Dictionary<ushort, Memory<byte>> CurrentXORResults = new Dictionary<ushort, Memory<byte>>();
+        private Dictionary<ushort, Memory<byte>> _currentXORResults = new Dictionary<ushort, Memory<byte>>();
 
         //Holds character to share object updates for
         public Entity entity;
-        public string entityName;
 
         //should be needed for transferring continents, probably can get away on just doing character changes? Could also be an easy way to flip character in and out of a channel on the client
         private bool _isActive = false;
+        public bool IsActive
+        {
+            get { return _isActive; }
+            set
+            {
+                if (value == _isActive)
+                    return;
+
+                else
+                {
+                    if (value == false)
+                    {
+                        _isActive = false;
+
+                        entity = null;
+
+                        DeactivateChannel();
+                        return;
+                    }
+
+                    _isActive = true;
+                }
+            }
+        }
 
         public ServerObjectUpdate(Session session, byte objectChannel)
         {
             _session = session;
-            ObjectChannel = objectChannel;
+            _objectChannel = objectChannel;
         }
 
-        public void AddObject(Entity entity1)
+        public void AddObject(Entity e)
         {
-            entity = entity1;
-            entityName = entity.CharName;
-            _isActive = true;
+            entity = e;
+            IsActive = true;
         }
 
         public void GenerateUpdate()
         {
-            if (entity == null)
-                return;
-
-            //See if character and current message has changed
-            if (!CompareObjects(baseXOR.Slice(1, 0xC8), entity.ObjectUpdate))
+            if (IsActive)
             {
-                Memory<byte> temp = new Memory<byte>(new byte[0xC9]);
-                temp.Span[0] = _isActive & (baseXOR.Span[0] == 0) ? (byte)1 : (byte)0;
-                CoordinateConversions.Xor_data(temp.Slice(1, 0xC8), entity.ObjectUpdate, baseXOR.Slice(1, 0xC8), 0xC8);
-                CurrentXORResults.Add(messageCounter, temp);
-                SessionQueueMessages.PackMessage(_session, temp, ObjectChannel);
+                //If entity becomes null, disable channel?
+                if (entity == null)
+                {
+                    IsActive = false;
+                    return;
+                }
+
+                //See if character and current message has changed
+                if (!CompareObjects(_baseXOR.Slice(1, 0xC8), entity.ObjectUpdate))
+                {
+                    Memory<byte> temp = new Memory<byte>(new byte[0xC9]);
+                    temp.Span[0] = _isActive & (_baseXOR.Span[0] == 0) ? (byte)1 : (byte)0;
+                    CoordinateConversions.Xor_data(temp.Slice(1, 0xC8), entity.ObjectUpdate, _baseXOR.Slice(1, 0xC8), 0xC8);
+                    _currentXORResults.Add(MessageCounter, temp);
+                    _session.sessionQueue.Add(new Message((MessageType)_objectChannel, MessageCounter, _baseMessageCounter == 0 ? (byte)0 : (byte)(MessageCounter - _baseMessageCounter), temp));
+                    MessageCounter++;
+                }
             }
         }
 
-        public void updateCharacter(Entity entity1)
+        private void DeactivateChannel()
         {
-            entity = entity1;
-            _isActive = true;
-        }
+            //No need to verify if entity is null or not, disabling channel anyway
+            Memory<byte> temp = new Memory<byte>(new byte[0xC9]);
 
-        public void DisableChannel()
-        {
-            entity = null;
-            if (_isActive)
-            {
-                _isActive = false;
-                Memory<byte> temp = new Memory<byte>(new byte[0xC9]);
-                temp.Span[0] = 1;
-                baseXOR.Slice(1, 0XC8).CopyTo(temp.Slice(1, 0xC8));
-                CurrentXORResults.Add(messageCounter, temp);
-                SessionQueueMessages.PackMessage(_session, temp, ObjectChannel);
-            }
+            //Since we are deactivating the channel, all we need to do is modify the first byte
+            temp.Span[0] = (_baseXOR.Span[0] == 1) ? (byte)1 : (byte)0;
+
+            _currentXORResults.Add(MessageCounter, temp);
+            _session.sessionQueue.Add(new Message((MessageType)_objectChannel, MessageCounter, _baseMessageCounter == 0 ? (byte)0 : (byte)(MessageCounter - _baseMessageCounter), temp));
+            MessageCounter++;
         }
 
         public void UpdateBaseXor(ushort msgCounter)
         {
             //Get the message client ack'd
-            Memory<byte> tempMemory = CurrentXORResults.GetValueOrDefault(msgCounter);
+            Memory<byte> tempMemory = _currentXORResults.GetValueOrDefault(msgCounter);
 
             if (tempMemory.Length == 0)
                 return;
 
             //xor base against this message
-            CoordinateConversions.Xor_data(baseXOR, tempMemory, 0xC9);
+            CoordinateConversions.Xor_data(_baseXOR, tempMemory, 0xC9);
 
             //Clear Dictionary
-            CurrentXORResults.Clear();
+            _currentXORResults.Clear();
 
             //Ensure this is new base
-            baseMessageCounter = msgCounter;
+            BaseMessageCounter = msgCounter;
         }
 
         private static bool CompareObjects(Memory<byte> first, Memory<byte> second)
@@ -109,7 +153,7 @@ namespace ReturnHome.Server.Network
             Span<byte> firstTemp = first.Span;
             Span<byte> secondTemp = second.Span;
 
-            for( int i = 0; i < first.Length; i++)
+            for (int i = 0; i < first.Length; i++)
             {
                 if (firstTemp[i] == secondTemp[i])
                     continue;
