@@ -8,13 +8,28 @@ namespace ReturnHome.Server.Network
 {
     public class SessionConnectionData
     {
-        public ushort lastReceivedMessageSequence { get; set; }
+        private Session _session;
+        private ushort _lastReceivedMessageSequence { get; set; } = 0;
         public ushort lastReceivedPacketSequence { get; set; }
         public ushort lastSentMessageSequence { get; set; }
         public ushort lastSentPacketSequence { get; set; }
         public ushort clientLastReceivedMessage { get; set; }
         public ushort clientLastReceivedMessageFinal { get; set; }
-        public ClientObjectUpdate client {get; set;}
+
+        //This property will trigger the rdpreport bool when it increments
+        public ushort lastReceivedMessageSequence
+        {
+            get { return _lastReceivedMessageSequence; }
+            set
+            {
+                if (_lastReceivedMessageSequence <= value)
+                {
+                    _lastReceivedMessageSequence = value;
+                    _session.PacketBodyFlags.RdpReport = true;
+                }
+            }
+        }
+        public ClientObjectUpdate client { get; set; }
 
         //Create an array of all 24 object updates
         //Need a way to ensure no more then 24 objects can be allocated, array may be best?
@@ -28,9 +43,9 @@ namespace ReturnHome.Server.Network
          * First theory is that this reset method should reside within the World Property for entity, 
          * checking if entity is a player first then calling this method to reset all channels if world changes
          */
-        public SessionConnectionData(Session _session)
+        public SessionConnectionData(Session session)
         {
-            lastReceivedMessageSequence = 0;
+            _session = session;
             lastReceivedPacketSequence = 0;
 
             //Client last soft ack of message's received.
@@ -40,8 +55,8 @@ namespace ReturnHome.Server.Network
             //and if client ack is higher, server will cycle through resend list and remove.
             clientLastReceivedMessageFinal = 0;
 
-			lastSentMessageSequence = 1;
-			lastSentPacketSequence = 1;
+            lastSentMessageSequence = 1;
+            lastSentPacketSequence = 1;
 
             //Create the object to track incoming client updates
             client = new();
@@ -55,32 +70,40 @@ namespace ReturnHome.Server.Network
 
         public void AddChannelObjects(List<Entity> charList)
         {
+            Span<ServerObjectUpdate> temp = serverObjects.Span;
+            charList.Remove(_session.MyCharacter);
             if (charList.Count == 0)
+            {
+                for (int i = 1; i < serverObjects.Length; i++)
+                {
+                    temp[i].IsActive = false;
+                }
                 return;
+            }
 
             charList = charList.GetRange(0, charList.Count > 23 ? 23 : charList.Count);
 
-            Span<ServerObjectUpdate> temp = serverObjects.Span;
             //Iterate over List from QuadTree against Channels
             for (int i = 1; i < serverObjects.Length; i++)
             {
                 //Character is already in a channel
-                if (charList.Contains(temp[i].entity))
+                if (charList.Contains(temp[i].entity) & temp[i].IsActive)
                 {
                     charList.Remove(temp[i].entity);
                     continue;
                 }
 
                 else
-                    temp[i].DisableChannel();
+                    //Calls the disable method for the channel
+                    temp[i].IsActive = false;
             }
 
             if (charList.Count == 0)
                 return;
 
-            for(int i = 0; i < serverObjects.Length; i++)
+            for (int i = 1; i < serverObjects.Length; i++)
             {
-                if(temp[i].entity == null)
+                if (!temp[i].IsActive)
                 {
                     temp[i].AddObject(charList[0]);
                     charList.RemoveAt(0);
@@ -88,6 +111,16 @@ namespace ReturnHome.Server.Network
                         return;
                 }
                 continue;
+            }
+        }
+
+        public void DisableChannels()
+        {
+            Span<ServerObjectUpdate> temp = serverObjects.Span;
+
+            for (int i = 0; i < serverObjects.Length; i++)
+            {
+                temp[i].IsActive = false;
             }
         }
     }
